+++ src/withClauseAttributes.ts
@@ -5,6 +5,7 @@ export const withClauseAttrCompletions: vscode.CompletionItem[] = [
   { name: "unwrap", default: 0 },
   { name: "recursionmode", default: 0 },
   { name: "spikemode", default: 0 },
+  { name: "iterationmode", default: 0 },
 ].map((attr) => {
   const item = new vscode.CompletionItem(`@${attr.name} ${attr.default}`, vscode.CompletionItemKind.Field);
   item.insertText = new vscode.SnippetString(`${attr.name} `).appendVariable("1", `${attr.default}`);