+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Include `iterationmode` in `for` loop intellisense.
+
 ### Removed
 
 ## [0.1.14] - 14-01-2025